Address PR review feedback
- Subscribe to authService via useSelector instead of getSnapshot reads
- Dispatch social.loginToggled from the sign-message callback; drop
  walletReauth state and the dispatching effect
- Use hasOwn check for URL error codes to avoid inherited-prop matches
- Extract ContentComponentProps + SettingMenuId to a leaf types module
  so settings-menu content components don't cycle through GameOptions
- Surface GOOGLE_LOGIN_DISABLED errorCode from 401 bodies in
  loadSession/autosave so active Google sessions land on the dedicated
  screen instead of generic SessionExpired
